tests/wpt/web-platform-tests/bluetooth/server/getPrimaryServices/gen-disconnect-called-during-error.https.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'disconnect() called during a getPrimaryServices ' +
    'call that fails. Reject with NetworkError.';
const expected = new DOMException(
    'GATT Server is disconnected. Cannot retrieve services. (Re)connect ' +
    'first with `device.gatt.connect`.', 'NetworkError');
let device;

bluetooth_test(() => getEmptyHealthThermometerDevice()
    .then(_ => ({device} = _))
    .then(() => {
      let promise = assert_promise_rejects_with_message(
        device.gatt.getPrimaryServices(),
        expected)
      device.gatt.disconnect();
      return promise;
    }),
    test_desc);

</script>
